+++ src/core/core-definitions.ts
@@ -41,6 +41,4 @@ export interface CliArgs extends Record<string, string | undefined> {
   service: string;
   serviceConfig?: string;
   matcher: string;
-  keySearch: string;
-  keyReplace: string;
 }