+++ src/file-formats/file-format-definitions.ts
@@ -14,8 +14,6 @@ export interface ReadTFileArgs {
   path: string;
   lng: string;
   format: TFileType;
-  keySearch: string;
-  keyReplace: string;
 }
 
 export interface TFileFormat {